Refactor first draft
