+++ project2/shared-libraries/daos/src/main/java/com/revature/daos/BusinessCardDAO.java
@@ -1,6 +1,6 @@
-package com.revature.businessCard.dao;
+package com.revature.daos;
 
-import com.revature.businessCard.model.BusinessCard;
+import com.revature.model.BusinessCard;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
