+++ project2/shared-libraries/daos/src/main/java/com/revature/daos/PersonalCardDAO.java
@@ -1,6 +1,6 @@
-package com.revature.personalCard.dao;
+package com.revature.daos;
 
-import com.revature.businessCard.model.BusinessCard;
+import com.revature.model.BusinessCard;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface PersonalCardDAO extends JpaRepository<BusinessCard, Integer> {
